refactor

+++ game/systems/animation_system.cpp
@@ -147,7 +147,7 @@ void animation_system::progress_animation_states(fixed_step& step) {
 
 			if (animation.frames.empty()) continue;
 
-			animation_state.player_position_ms += delta.in_milliseconds() * animation_state.speed_factor;
+			animation_state.player_position_ms += static_cast<float>(delta.in_milliseconds()) * animation_state.speed_factor;
 
 			while (true) {
 				float frame_duration = animation.frames[animation_state.get_current_frame()].duration_milliseconds;
